+++ packages/eslint-plugin/CHANGELOG.md
@@ -7,6 +7,10 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## Unreleased -->
 
+### Changed
+
+- Fix conflicting `@typescript-eslint/naming-convention` rules  [[#247](https://github.com/Shopify/web-configs/pull/247)]
+
 ## 40.2.2 - 2021-05-05
 
 ### Changed